service_2/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
requests
